feat(all-about-som): server - 현재 위치 조회 api 추가
all-about-som
